Change the feteching of user info from useEffect to useSWR and clean up states
