+++ leetnode/src/components/settings/Calendar.tsx
@@ -1,7 +1,8 @@
 export default function Calendar() {
   return (
     <>
-      <h1>Calender</h1>
+      <h1 className="text-center">Calendar</h1>
+      <hr className="h-px my-4 bg-gray-200 border-0" />
     </>
   );
 }
